Move mark_as_read to discussions controller
rip out discussion reader api pipeline
