+++ app/controllers/api/discussions_controller.rb
@@ -1,21 +1,20 @@
 class API::DiscussionsController < API::RestfulController
   load_and_authorize_resource only: [:show, :mark_as_read, :set_volume]
-  load_resource only: [:create, :update]
+  load_resource only: [:create, :update, :mark_as_read]
 
   def dashboard
     instantiate_collection { |collection| collection_for_dashboard collection }
-    respond_with_collection serializer: DiscussionWrapperSerializer, root: 'discussion_wrappers'
+    respond_with_collection
   end
 
   def inbox
     instantiate_collection { |collection| collection_for_inbox collection }
-    respond_with_collection serializer: DiscussionWrapperSerializer, root: 'discussion_wrappers'
+    respond_with_collection
   end
 
-  def index
-    load_and_authorize :group if params[:group_id] || params[:group_key]
-    instantiate_collection
-    respond_with_collection serializer: DiscussionWrapperSerializer, root: 'discussion_wrappers'
+  def mark_as_read
+    DiscussionReader.for(user: current_user, discussion: resource).viewed! (discussion_event || resource).created_at
+    respond_with_resource
   end
 
   private
@@ -25,6 +24,7 @@ class API::DiscussionsController < API::RestfulController
   end
 
   def visible_groups
+    load_and_authorize :group if params[:group_id] || params[:group_key]
     Array(@group).presence || current_user.groups
   end
 
@@ -40,7 +40,8 @@ class API::DiscussionsController < API::RestfulController
     collection.not_muted.unread.sorted_by_latest_activity
   end
 
-  def collection=(value)
-    @discussions = DiscussionWrapper.new_collection user: current_user, discussions: value
+  def discussion_event
+    Event.where(discussion_id: resource.id, sequence_id: params[:sequence_id]).first
   end
+
 end
